add sp transformation

+++ src/NosCore.GameObject/MapNpc.cs
@@ -44,9 +44,9 @@ namespace NosCore.GameObject
         public byte Speed { get; set; }
         public int Mp { get; set; }
         public int Hp { get; set; }
-        public byte Morph { get; set; }
+        public short Morph { get; set; }
         public byte MorphUpgrade { get; set; }
-        public byte MorphDesign { get; set; }
+        public short MorphDesign { get; set; }
         public byte MorphBonus { get; set; }
         public bool NoAttack { get; set; }
         public bool NoMove { get; set; }
